Tests/LibWeb/Text/input/Streams/ReadableStream-pipeTo-cannot-pipe-to-locked-stream.html

<script src="../include.js"></script>
<script>
    asyncTest(done => {
        const writableStream = new WritableStream(
            {
                write(chunk) {}
            }
        );
        const stream = new ReadableStream({
            start(controller) {},
            pull(controller) {},
            cancel() {},
        });

        writableStream.getWriter();
        stream.pipeTo(writableStream)
            .catch(reason => {
                println(reason);
                done();
            });
    });
</script>
