Tests/LibWeb/Text/expected/Streams/ReadableStream-pipeTo-cannot-pipe-to-locked-stream.txt

TypeError: Failed to execute 'pipeTo' on 'ReadableStream':  Cannot pipe to a locked stream
